requirements/base.txt
numpy
pillow
pydantic
fastapi
uvicorn[standard]
python-multipart
PyYAML
